Fix integers. Also fix bad cast that forced FAT12.

+++ cherios/system/fatfs/include/integer.h
@@ -22,16 +22,16 @@ typedef unsigned int	UINT;
 typedef unsigned char	BYTE;
 
 /* These types MUST be 16-bit */
-typedef short			SHORT;
-typedef unsigned short	WORD;
-typedef unsigned short	WCHAR;
+typedef int16_t			SHORT;
+typedef uint16_t	    WORD;
+typedef uint16_t	    WCHAR;
 
 /* These types MUST be 32-bit */
-typedef long			LONG;
-typedef unsigned long	DWORD;
+typedef int32_t			LONG;
+typedef uint32_t	    DWORD;
 
 /* This type MUST be 64-bit (Remove this for C89 compatibility) */
-typedef unsigned long long QWORD;
+typedef uint64_t        QWORD;
 
 #endif
